+++ packages/web3-eth/test/integration/eth.test.ts
@@ -22,7 +22,6 @@ import { Contract } from 'web3-eth-contract';
 import { SupportedProviders } from 'web3-core';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import IpcProvider from 'web3-providers-ipc';
-import { hexToNumber } from 'web3-utils';
 import { Web3Eth } from '../../src';
 
 import {
@@ -113,31 +112,6 @@ describe('eth', () => {
 		it('givenProvider', () => {
 			const { givenProvider } = web3Eth;
 			expect(givenProvider).toBeUndefined();
-		});
-		it('BatchRequest', async () => {
-			const batch = new web3Eth.BatchRequest();
-			const request1 = {
-				id: 10,
-				method: 'eth_getBalance',
-				params: [accounts[0], 'latest'],
-			};
-			const request2 = {
-				id: 11,
-				method: 'eth_getBalance',
-				params: [accounts[1], 'latest'],
-			};
-			batch.add(request1).catch(console.error);
-			batch.add(request2).catch(console.error);
-			const [response1, response2] = await batch.execute();
-			// eslint-disable-next-line jest/no-standalone-expect
-			expect(response1.result).toBeDefined();
-			// eslint-disable-next-line jest/no-standalone-expect
-			expect(response2.result).toBeDefined();
-			// TODO: in future release add test for validation of returned results , ( match balance )
-			// eslint-disable-next-line jest/no-standalone-expect
-			expect(Number(hexToNumber(String(response1.result)))).toBeGreaterThan(0);
-			// eslint-disable-next-line jest/no-standalone-expect
-			expect(Number(hexToNumber(String(response2.result)))).toBeGreaterThan(0);
 		});
 	});
 });
