+++ packages/web3-eth/test/integration/helper.ts
@@ -14,8 +14,10 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { Web3PromiEvent } from 'web3-common';
-import { ReceiptInfo, SendTransactionEvents, Web3Eth } from '../../src';
+import { Block, FMT_NUMBER, Web3PromiEvent } from 'web3-common';
+import { AbiEventFragment } from 'web3-eth-abi';
+import { ReceiptInfo, SendTransactionEvents, TransactionInfo, Web3Eth } from '../../src';
+import { BasicAbi } from '../shared_fixtures/build/Basic';
 
 type SendFewTxParams = {
 	web3Eth: Web3Eth;
@@ -53,3 +55,72 @@ export const sendFewTxes = async ({
 	}
 	return res;
 };
+
+export const validateTransaction = (tx: TransactionInfo) => {
+	expect(tx.nonce).toBeDefined();
+	expect(tx.hash).toBeDefined();
+	expect(String(tx.hash)?.length).toBe(66);
+	expect(Number(tx.type)).toBe(0);
+	expect(tx.blockHash).toBeDefined();
+	expect(String(tx.blockHash)?.length).toBe(66);
+	expect(Number(tx.blockNumber)).toBeGreaterThan(0);
+	expect(tx.transactionIndex).toBeDefined();
+	expect(tx.from?.length).toBe(42);
+	expect(tx.to?.length).toBe(42);
+	expect(Number(tx.value)).toBe(1);
+	expect(tx.input).toBe('0x');
+	expect(tx.r).toBeDefined();
+	expect(tx.s).toBeDefined();
+	expect(Number(tx.gas)).toBeGreaterThan(0);
+};
+export const validateBlock = (b: Block) => {
+	expect(b.nonce).toBeDefined();
+	expect(Number(b.baseFeePerGas)).toBeGreaterThan(0);
+	expect(b.number).toBeDefined();
+	expect(b.hash).toBeDefined();
+	expect(b.parentHash?.length).toBe(66);
+	expect(b.sha3Uncles?.length).toBe(66);
+	expect(b.transactionsRoot).toHaveLength(66);
+	expect(b.receiptsRoot).toHaveLength(66);
+	expect(b.logsBloom).toBeDefined();
+	expect(b.miner).toHaveLength(42);
+	expect(b.difficulty).toBeDefined();
+	expect(b.stateRoot).toHaveLength(66);
+	expect(b.gasLimit).toBeDefined();
+	expect(b.gasUsed).toBeDefined();
+	expect(b.timestamp).toBeDefined();
+	expect(b.extraData).toBeDefined();
+	expect(b.mixHash).toBeDefined();
+	expect(b.totalDifficulty).toBeDefined();
+	expect(b.baseFeePerGas).toBeDefined();
+	expect(b.size).toBeDefined();
+	expect(Array.isArray(b.transactions)).toBe(true);
+	expect(Array.isArray(b.uncles)).toBe(true);
+};
+export const validateReceipt = (r: ReceiptInfo) => {
+	expect(r.transactionHash).toBeDefined();
+	expect(r.transactionIndex).toBeDefined();
+	expect(r.blockHash).toBeDefined();
+	expect(r.blockNumber).toBeDefined();
+	expect(r.from).toBeDefined();
+	expect(r.to).toBeDefined();
+	expect(r.cumulativeGasUsed).toBeDefined();
+	expect(r.gasUsed).toBeDefined();
+	expect(r.effectiveGasPrice).toBeDefined();
+	expect(r.logs).toBeDefined();
+	expect(r.logsBloom).toBeDefined();
+	expect(r.status).toBeDefined();
+	expect(String(r.transactionHash)).toHaveLength(66);
+	expect(Number(r.gasUsed)).toBeGreaterThan(0);
+};
+
+export const mapFormatToType: { [key: string]: string } = {
+	[FMT_NUMBER.NUMBER]: 'number',
+	[FMT_NUMBER.HEX]: 'string',
+	[FMT_NUMBER.STR]: 'string',
+	[FMT_NUMBER.BIGINT]: 'bigint',
+};
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+export const eventAbi: AbiEventFragment = BasicAbi.find((e: any) => {
+	return e.name === 'StringEvent' && (e as AbiEventFragment).type === 'event';
+})! as AbiEventFragment;
