+++ packages/web3-eth/test/integration/setup.js
@@ -19,6 +19,6 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 require('../config/setup');
 
-const jestTimeout = 35000;
+const jestTimeout = 15000;
 
 jest.setTimeout(jestTimeout);
